app/requirements.txt
flask
structlog[dev]
python-logstash
twisted